requirements.txt
fastapi
uvicorn[standard]
transformers
torch>=2.5.1
chromadb
pydantic
numpy
scipy
python-multipart
PyPDF2
python-docx
openpyxl